third_party/WebKit/LayoutTests/editing/execCommand/inserting-ordered-list-crash.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
// This is a test for crash bag for https://crbug.com/343799.
test(()=> assert_selection(
  [
  "<div contenteditable>",
    "<strike>",
      "^<input>",
    "</strike>",
    "<hr>",
    "<strike>",
      ">|",
    "</strike>",
  "</div>"
  ] ,
  'InsertOrderedList',
  [
  "<div contenteditable>",
    "<ol>",
      "<li>",
        "<strike>",
          "^<input>",
        "</strike>",
      "</li>",
      "<li>",
        "<hr>|",
      "</li>",
      "<li>",
        "<strike>",
          ">",
        "</strike>",
      "</li>",
    "</ol>",
  "</div>"
  ]),
  "This InsertOrderedList should not cause page crash.");
</script>
